+++ tests/conftest.py
@@ -41,7 +41,7 @@ from stac_fastapi.extensions.core.sort import SortConformanceClasses
 from stac_fastapi.extensions.third_party import BulkTransactionExtension
 from stac_pydantic import Collection, Item
 
-from stac_fastapi.pgstac.config import Settings
+from stac_fastapi.pgstac.config import PostgresSettings, Settings
 from stac_fastapi.pgstac.core import CoreCrudClient
 from stac_fastapi.pgstac.db import close_db_connection, connect_to_db
 from stac_fastapi.pgstac.extensions import QueryExtension
@@ -111,18 +111,12 @@ async def pgstac(database):
     ],
     scope="session",
 )
-def api_client(request, database):
+def api_client(request):
     hydrate, prefix, response_model = request.param
     api_settings = Settings(
-        postgres_user=database.user,
-        postgres_pass=database.password,
-        postgres_host_reader=database.host,
-        postgres_host_writer=database.host,
-        postgres_port=database.port,
-        postgres_dbname=database.dbname,
-        use_api_hydrate=hydrate,
         enable_response_models=response_model,
         testing=True,
+        use_api_hydrate=hydrate,
     )
 
     api_settings.openapi_url = prefix + api_settings.openapi_url
@@ -203,11 +197,19 @@ def api_client(request, database):
 
 
 @pytest.fixture(scope="function")
-async def app(api_client):
+async def app(api_client, database):
+    postgres_settings = PostgresSettings(
+        postgres_user=database.user,
+        postgres_pass=database.password,
+        postgres_host_reader=database.host,
+        postgres_host_writer=database.host,
+        postgres_port=database.port,
+        postgres_dbname=database.dbname,
+    )
     logger.info("Creating app Fixture")
     time.time()
     app = api_client.app
-    await connect_to_db(app)
+    await connect_to_db(app, postgres_settings=postgres_settings)
 
     yield app
 
@@ -290,14 +292,8 @@ async def load_test2_item(app_client, load_test_data, load_test2_collection):
 @pytest.fixture(
     scope="session",
 )
-def api_client_no_ext(database):
+def api_client_no_ext():
     api_settings = Settings(
-        postgres_user=database.user,
-        postgres_pass=database.password,
-        postgres_host_reader=database.host,
-        postgres_host_writer=database.host,
-        postgres_port=database.port,
-        postgres_dbname=database.dbname,
         testing=True,
     )
     return StacApi(
@@ -310,11 +306,19 @@ def api_client_no_ext(database):
 
 
 @pytest.fixture(scope="function")
-async def app_no_ext(api_client_no_ext):
+async def app_no_ext(api_client_no_ext, database):
+    postgres_settings = PostgresSettings(
+        postgres_user=database.user,
+        postgres_pass=database.password,
+        postgres_host_reader=database.host,
+        postgres_host_writer=database.host,
+        postgres_port=database.port,
+        postgres_dbname=database.dbname,
+    )
     logger.info("Creating app Fixture")
     time.time()
     app = api_client_no_ext.app
-    await connect_to_db(app)
+    await connect_to_db(app, postgres_settings=postgres_settings)
 
     yield app